OAK-1620 Index cost calculation integer overflow
git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1581892 13f79535-47bb-0310-9956-ffa450edef68

+++ oak-core/src/main/java/org/apache/jackrabbit/oak/plugins/index/property/OrderedPropertyIndexLookup.java
@@ -44,7 +44,7 @@ public class OrderedPropertyIndexLookup extends PropertyIndexLookup {
     /**
      * we're slightly more expensive than the standard PropertyIndex.
      */
-    private static final int COST_OVERHEAD = 3;
+    private static final double COST_OVERHEAD = 3;
     
     public OrderedPropertyIndexLookup(NodeState root) {
         super(root);
